Add alert

+++ src/components/DisplayComments.js
@@ -1,6 +1,7 @@
 import React, { useRef } from 'react'
 import { useDeleteCommentMutation, useEditCommentMutation } from '../features/commentAPI';
 import '../styles/Comment.css'
+import Swal from 'sweetalert2'
 
 export default function Comments(props) {
 
@@ -20,29 +21,53 @@ export default function Comments(props) {
         roleLocal = "admin"
     } else if(userLocal){
         roleLocal = "user"
-        idUserLocal = userLocal._id
+        idUserLocal = userLocal.id
     }
     
     const [deleteComment] = useDeleteCommentMutation()
     const deletingComment = () => {
-        deleteComment(id)
+        Swal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            confirmButtonColor: '#28a745',
+            cancelButtonColor: '#dd3544',
+            showCancelButton: true,
+            confirmButtonText: 'Yes, delete it!'
+        }).then((result) => {
+          if (result.isConfirmed) {
+              deleteComment(id)
+              Swal.fire(
+                  'Deleted!',
+                  'Your file has been deleted.',
+                  'success'
+                )
+            window.location.reload(false);
+        }})
     }
 
     const handleSubmit = async (e) => {
         e.preventDefault();
-        const {data, error} = await editComment({comment:commentRef.current.value, id: id});
-        if(data){
-            console.log(data)
-        } else {
-            console.log(error)
-        }
+        Swal.fire({
+            title: 'Do you want to save the changes?',
+            showDenyButton: true,
+            showCancelButton: true,
+            confirmButtonText: 'Save',
+            denyButtonText: `Don't save`,
+        }).then((result) => {
+            if (result.isConfirmed) {
+                editComment({comment:commentRef.current.value, id: id});
+                Swal.fire('Saved!', '', 'success')
+            } else if (result.isDenied) {
+                Swal.fire('Changes are not saved', '', 'info')
+            }
+        })
     }
 
   return (
     <div className='Comment-Container'>
         <img className='Comment-img' src={comment.user.photo} alt='User'/>
         <div className="container">
-            <div className="container_terminal"></div>
             <div className="terminal_toolbar">
                 <div className='buttons-name'>
                     <div className="butt">
@@ -60,20 +85,16 @@ export default function Comments(props) {
                         <path d="M10 10l4 4m0 -4l-4 4" />
                     </svg>
                 </>
-                : null
-                }
-
+                : null}
             </div>
 
             <div className="terminal_body">
                 {(userLocal && (roleLocal === "admin" || idUserLocal === commentUserId))?
                 <>
                     <form className="terminal_promt" onSubmit={e => handleSubmit(e)}>
-                        <div>
-                            <input id={id} className='comment-input' type="text" name="comment" required size="30"
-                                defaultValue={comment.comment}
-                                minLength="3" maxLength="350" ref={commentRef}/>
-                        </div>                    
+                        
+                        <input id={id} className='comment-input' type="text" name="comment" required defaultValue={comment.comment} minLength="3" maxLength="350" ref={commentRef}/>
+                                          
                         <button className='comment-edit-button' type='submit'>
                         <svg className="comment-edit" xmlns="http://www.w3.org/2000/svg"  width="32" height="32" viewBox="0 0 24 24" strokeWidth="1.5" stroke="#ffec00" fill="none" strokeLinecap="round" strokeLinejoin="round">
                         <path stroke="none" d="M0 0h24v24H0z" fill="none"/>
@@ -87,7 +108,6 @@ export default function Comments(props) {
                 :
                     <div className="terminal_promt">
                         <span className="terminal_bling">{comment.comment}</span>
-                        <span className="terminal_cursor"></span>
                     </div>
                 }
             </div>
